# perf(tokenizer): did some perf