**main** 
```bash
Gnuplot not found, using plotters backend
tokenizer/small(7K)     time:   [13.125 us 13.205 us 13.287 us]                                 
Found 12 outliers among 100 measurements (12.00%)
  3 (3.00%) high mild
  9 (9.00%) high severe

tokenizer/large(201K)   time:   [361.35 us 361.50 us 361.66 us]                                  
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

**current**
```bash
tokenizer/small(7K)     time:   [8.9518 us 8.9584 us 8.9651 us]                                 
                        change: [-32.092% -31.880% -31.688%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild

tokenizer/large(201K)   time:   [296.80 us 297.03 us 297.25 us]                                  
                        change: [-17.994% -17.868% -17.743%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  13 (13.00%) high mild
  1 (1.00%) high severe
```